refactor(RouteViewer): clean up, extract routeRow to new file
