+++ lib/components/map/connected-transit-vehicle-overlay.js
@@ -41,8 +41,11 @@ function VehicleTooltip (props) {
       : intl.formatMessage({
         id: `components.TransitVehicleOverlay.line`
       })
-    // Only render space if name is present
-    name = `${mode}${name && ' '}${name}`
+    name = intl.formatMessage(
+      { id: 'components.TransitVehicleOverlay.realtimeVehicleName' },
+      // Only render space if name is present
+      { headsign: `${name && ' '}${name}`, mode }
+    )
   }
 
   const stopStatus = vehicle.stopStatus || 'in_transit_to'
@@ -54,10 +57,7 @@ function VehicleTooltip (props) {
     <Tooltip direction={direction} permanent={permanent}>
       <span>
         <strong style={{ fontSize: '110%' }}>
-          <FormattedMessage
-            id='components.TransitVehicleOverlay.realtimeVehicleName'
-            values={{ name }}
-          />
+          {name}
         </strong>
         <FormattedMessage
           id='components.TransitVehicleOverlay.relativeTime'
